+++ tests/fixtures/controls/dummy_need_access_to_end_state.py
@@ -1,0 +1,9 @@
+from chaoslib.types import Experiment, Journal
+
+
+def after_hypothesis_control(context: Experiment, state: Journal, **kwargs):
+    state["after_hypothesis_control"] = True
+
+
+def after_experiment_control(context: Experiment, state: Journal, **kwargs):
+    state["after_experiment_control"] = True